main/memory: improve drawHeapTitle accumulator signedness (#5540)

+++ src/memory.cpp
@@ -1402,8 +1402,8 @@ checkHeapNode:
 void CMemory::CStage::drawHeapTitle(int y)
 {
     int node = (stageGetAllocationMode(this) == 2) ? stageGetHeapHead(this) : *reinterpret_cast<int*>(stageGetHeapHead(this) + 8);
-    int totalUnuse = 0;
-    int maxUnuse = 0;
+    unsigned int totalUnuse = 0;
+    unsigned int maxUnuse = 0;
     int prev = *reinterpret_cast<int*>(node + 4);
     char line[264];
     char* sourceName = reinterpret_cast<char*>(reinterpret_cast<unsigned char*>(this) + 0x10);
@@ -1414,16 +1414,16 @@ void CMemory::CStage::drawHeapTitle(int y)
             strcpy(line, sourceName + ((srcLen - 12U) & ~((srcLen - 12U) >> 31)));
             Graphic.DrawDebugStringDirect(0x10, static_cast<unsigned short>(y), line, 8);
             int maxRound = 0;
-            if ((maxUnuse < 0) && ((maxUnuse & 0x3ff) != 0)) {
+            if ((static_cast<int>(maxUnuse) < 0) && ((maxUnuse & 0x3ff) != 0)) {
                 maxRound = 1;
             }
             int totalRound = 0;
-            if ((totalUnuse < 0) && ((totalUnuse & 0x3ff) != 0)) {
+            if ((static_cast<int>(totalUnuse) < 0) && ((totalUnuse & 0x3ff) != 0)) {
                 totalRound = 1;
             }
 
             sprintf(line, s_drawHeapTitleFmt, *reinterpret_cast<int*>(reinterpret_cast<unsigned char*>(this) + 0x124),
-                    (totalUnuse >> 10) + totalRound, (maxUnuse >> 10) + maxRound);
+                    (static_cast<int>(totalUnuse) >> 10) + totalRound, (static_cast<int>(maxUnuse) >> 10) + maxRound);
             Graphic.DrawDebugStringDirect(0x208, y, line, 8);
             return;
         }
@@ -1431,7 +1431,7 @@ void CMemory::CStage::drawHeapTitle(int y)
         if ((*reinterpret_cast<unsigned char*>(node + 2) & 4) == 0) {
             int blockSize = *reinterpret_cast<int*>(node + 8) - (node + 0x40);
             totalUnuse += blockSize;
-            if (maxUnuse < blockSize) {
+            if (static_cast<int>(maxUnuse) < blockSize) {
                 maxUnuse = blockSize;
             }
         }
